third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-breakpoints/set-breakpoint-expected.txt

Tests setting breakpoints.


Running: testSetBreakpoint
Script source was shown.
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (set-breakpoint.html:14)
Breakpoint sidebar pane 
set-breakpoint.html:14  var x = Math.sqrt(10);
Breakpoint sidebar pane 
No breakpoints
Script execution resumed.

Running: testSetBreakpointOnTheLastLine
Call stack:
    0) oneLineTestFunction (set-breakpoint.html:4)

Running: testSetBreakpointOnTheLastLine2
Call stack:
    0) oneLineTestFunction2 (set-breakpoint.html:8)

Running: testSetBreakpointOnTheSameLine
